Numbers and start seq extraction
Reasonably happy with everything I've done to get numbers and information for the 3' UTRs that might be of use. Started to work on extracting sequences to have for analyses identifying specific mismatches as well as having them available to run in Profiler. Encountered bug where some Hawaii sequences could not be extracted due to the start being after the stop.